import of plot_histogram, more compact save and load methods, test for save and load

+++ quaccatoo/utils.py
@@ -2,7 +2,6 @@
 This module provides functions to save and load quaccatoo objects, such as instances from QSys, ExpData, Analysis and PulsedSim.
 """
 
-import inspect
 import os
 import shutil
 import zipfile
@@ -11,12 +10,12 @@ import dill
 import numpy as np
 from qutip import Qobj, fileio
 
-__all__ = ["save", "load"]
+__all__ = ["save_quaccatoo", "load_quaccatoo"]
 
 ####################################################################################################
 
 
-def save(obj, file_name):
+def save_quaccatoo(obj_save, file_path):
     """
     Look for all the attributes of the obj, save Qobj attributes to separate files,
     and save the rest of the attributes to a pickle file. Finally, create a zip file
@@ -24,33 +23,28 @@ def save(obj, file_name):
 
     Parameters
     ----------
-    obj : quaccatoo obj
+    obj_save : quaccatoo obj
         The obj to be saved.
-    file_name : str
+    file_path : str
         Path to the file where the attributes will be saved.
     """
-    if not isinstance(file_name, str):
-        raise ValueError("file_name must be a string")
+    if not isinstance(file_path, str):
+        raise ValueError("file_path must be a string")
 
-    # create a temporary directory to store files
     tmp_dir = "tmp"
-    os.makedirs(tmp_dir, exist_ok=True)
 
     try:
-        # Get a list of all the attributes defined in the obj
-        attributes = [
-            attr
-            for attr in dir(obj)
-            if not attr.startswith("__")
-            and not (inspect.isfunction(getattr(obj, attr)) or inspect.ismethod(getattr(obj, attr)))
-        ]
+        # create a temporary directory to store files and get a list of all the attributes defined in the obj
+        os.makedirs(tmp_dir, exist_ok=True)
+        attributes = list(obj_save.__dict__.keys())
         py_attr = []
 
         # Separate attributes into Python and Qobj types
         for attr in attributes:
-            value = getattr(obj, attr)
+            value = getattr(obj_save, attr)
             if isinstance(value, Qobj) or (
                 isinstance(value, (list, np.ndarray))
+                and len(value) > 0
                 and all(isinstance(item, Qobj) for item in value)
             ):
                 # Save Qobj attributes to a file in the temporary directory
@@ -60,8 +54,8 @@ def save(obj, file_name):
 
         # Create a dictionary to store the python attributes names and values
         py_data = {
-            "__type__": obj.__class__.__name__,
-            **{attr: getattr(obj, attr) for attr in py_attr},
+            "__type__": obj_save.__class__.__name__,
+            **{attr: getattr(obj_save, attr) for attr in py_attr},
         }
 
         # Save the python data to a file in the temporary directory
@@ -69,7 +63,7 @@ def save(obj, file_name):
             dill.dump(py_data, f_pkl)
 
         # Create a zip file to store all the files
-        with zipfile.ZipFile(file_name, "w") as zip_file:
+        with zipfile.ZipFile(file_path, "w") as zip_file:
             for root, _, files in os.walk(tmp_dir):
                 for file in files:
                     file_path = os.path.join(root, file)
@@ -77,18 +71,18 @@ def save(obj, file_name):
                     zip_file.write(file_path, rel_path)
 
     finally:
-        # Remove the temporary directory
-        shutil.rmtree(tmp_dir)
+        if os.path.exists(tmp_dir):
+            shutil.rmtree(tmp_dir)
 
 
-def load(file_name):
+def load_quaccatoo(file_path):
     """
     Loads the attributes of an obj from a zip file,
     creates an instance of the obj, and sets the attributes.
 
     Parameters
     ----------
-    file_name : str
+    file_path : str
         Path to the zip file where the attributes are saved.
 
     Returns
@@ -96,42 +90,40 @@ def load(file_name):
     obj
         The loaded obj.
     """
-    if not isinstance(file_name, str):
+    if not isinstance(file_path, str):
         raise ValueError("file_name must be a string")
 
-    # Extract the zip file to a temporary directory
     tmp_dir = "tmp"
 
     import quaccatoo  # pylint: disable=import-outside-toplevel,cyclic-import
 
     try:
+        # Extract the zip file to a temporary directory
         os.makedirs(tmp_dir, exist_ok=True)
-        with zipfile.ZipFile(file_name, "r") as zip_file:
+        with zipfile.ZipFile(file_path, "r") as zip_file:
             zip_file.extractall(tmp_dir)
 
         # Load the py_data.pkl file
         with open(os.path.join(tmp_dir, "py_data.pkl"), "rb") as f_pkl:
             py_data = dill.load(f_pkl)
 
-        # Get the obj name from the py_data
+        # Get the obj name from the py_data and create new instance
         cls = getattr(quaccatoo, py_data["__type__"])
-
-        # Create an empty instance
-        obj = cls.__new__(cls)
+        obj_load = cls.__new__(cls)
 
         # Load the attributes from the py_data
         for attr, value in py_data.items():
             if attr != "__type__":
-                setattr(obj, attr, value)
+                setattr(obj_load, attr, value)
 
         # Load the Qobj attributes from the files
         for file in os.listdir(tmp_dir):
             if file != "py_data.pkl":
                 attr_name = os.path.splitext(file)[0]
                 value = fileio.qload(os.path.join(tmp_dir, attr_name))
-                setattr(obj, attr_name, value)
+                setattr(obj_load, attr_name, value)
 
-        return obj
+        return obj_load
 
     finally:
         if os.path.exists(tmp_dir):
